ftables changes

+++ database/migrations/2021_08_15_093747_create_vendors_table.php
@@ -17,7 +17,7 @@ class CreateVendorsTable extends Migration
             $table->id();
             $table->string('name');
             $table->integer('item_id');
-            $table->float('price');
+            $table->BigInteger('price');
             $table->timestamps();
         });
     }
